Sandy bot/requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.0.0
python-docx>=0.8.0
notion-client>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
pywin32>=300; sys_platform == 'win32'
